requirements (1).txt
requests==2.31.0
pyyaml==6.0.1
scikit-learn==1.4.2
jinja2==3.1.2
reportlab==4.2.2
python-nmap==0.7.1
selenium==4.22.0
langchain==0.2.7
transformers==4.42.4
webdriver-manager==4.0.2